Get Copy Number Sequence Similarity or Distance Matrix

Get Copy Number Sequence Similarity or Distance Matrix

sh_get_score_matrix(
  x,
  sub_mat = NULL,
  simple_version = FALSE,
  block_size = NULL,
  dislike = FALSE,
  cores = 1L,
  verbose = FALSE
)

Arguments

x
a coding copy number sequence (valid letters are A to X).

sub_mat
a substitution matrix, each element indicates a score to plus. See sh_build_sub_matrix().

simple_version
if TRUE, just use segmental copy number value.

block_size
a block size to aggregrate, this is designed for big data, it means results from adjacent sequences will be aggregrate by means to reduce the size of result matrix.

dislike
if TRUE, returns a dissimilarity matrix instead of a similarity matrix.

cores
computer cores, default is 1, note it is super fast already, set more cores typically do not speed up the computation.

verbose
if TRUE, print extra message, note it will slower the computation.

Value

a score matrix.

Examples

load(system.file("extdata", "toy_segTab.RData",
  package = "sigminer.helper", mustWork = TRUE
))
x <- sh_coding_segs(segTabs)
x

#> $dt
#>                            sample chromosome     start       end segVal lenVal
#>   1: TCGA-05-4417-01A-22D-1854-01          1   3218923 116319008      2      4
#>   2: TCGA-05-4417-01A-22D-1854-01          1 116324707 120523902      1      3
#>   3: TCGA-05-4417-01A-22D-1854-01          1 149879545 161395555      4      4
#>   4: TCGA-05-4417-01A-22D-1854-01          1 161398414 247812431      4      4
#>   5: TCGA-05-4417-01A-22D-1854-01          2    499482 117951931      3      4
#>  ---                                                                          
#> 463: TCGA-DF-A2KN-01A-11D-A17U-01         22  17423930  29440692      2      4
#> 464: TCGA-DF-A2KN-01A-11D-A17U-01         22  29442666  29805407      3      2
#> 465: TCGA-DF-A2KN-01A-11D-A17U-01         22  29824523  47176707      2      4
#> 466: TCGA-DF-A2KN-01A-11D-A17U-01         22  47184783  47249551      2      2
#> 467: TCGA-DF-A2KN-01A-11D-A17U-01         22  47258179  49331012      2      3
#>                                   ID Seqs
#>   1:  TCGA-05-4417-01A-22D-1854-01:1    L
#>   2:  TCGA-05-4417-01A-22D-1854-01:1    G
#>   3:  TCGA-05-4417-01A-22D-1854-01:1    T
#>   4:  TCGA-05-4417-01A-22D-1854-01:1    T
#>   5:  TCGA-05-4417-01A-22D-1854-01:2    P
#>  ---                                     
#> 463: TCGA-DF-A2KN-01A-11D-A17U-01:22    L
#> 464: TCGA-DF-A2KN-01A-11D-A17U-01:22    N
#> 465: TCGA-DF-A2KN-01A-11D-A17U-01:22    L
#> 466: TCGA-DF-A2KN-01A-11D-A17U-01:22    J
#> 467: TCGA-DF-A2KN-01A-11D-A17U-01:22    K
#> 
#> $map
#>  10  20  30  40  11  21  31  41  12  22  32  42  13  23  33  43  14  24  34  44 
#> "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L" "M" "N" "O" "P" "Q" "R" "S" "T" 
#>  15  25  35  45 
#> "U" "V" "W" "X" 
#> 
#> $mat
#>     A   B   C   D   E   F   G   H   I   J   K   L   M   N   O   P   Q   R   S
#> A  24  18  12   6  20  15  10   5  -8  -6  -4  -2 -12  -9  -6  -3 -16 -12  -8
#> B  18  24  18  12  15  20  15  10  -6  -8  -6  -4  -9 -12  -9  -6 -12 -16 -12
#> C  12  18  24  18  10  15  20  15  -4  -6  -8  -6  -6  -9 -12  -9  -8 -12 -16
#> D   6  12  18  24   5  10  15  20  -2  -4  -6  -8  -3  -6  -9 -12  -4  -8 -12
#> E  20  15  10   5  24  18  12   6  -4  -3  -2  -1  -8  -6  -4  -2 -12  -9  -6
#> F  15  20  15  10  18  24  18  12  -3  -4  -3  -2  -6  -8  -6  -4  -9 -12  -9
#> G  10  15  20  15  12  18  24  18  -2  -3  -4  -3  -4  -6  -8  -6  -6  -9 -12
#> H   5  10  15  20   6  12  18  24  -1  -2  -3  -4  -2  -4  -6  -8  -3  -6  -9
#> I  -8  -6  -4  -2  -4  -3  -2  -1  24  18  12   6  -4  -3  -2  -1  -8  -6  -4
#> J  -6  -8  -6  -4  -3  -4  -3  -2  18  24  18  12  -3  -4  -3  -2  -6  -8  -6
#> K  -4  -6  -8  -6  -2  -3  -4  -3  12  18  24  18  -2  -3  -4  -3  -4  -6  -8
#> L  -2  -4  -6  -8  -1  -2  -3  -4   6  12  18  24  -1  -2  -3  -4  -2  -4  -6
#> M -12  -9  -6  -3  -8  -6  -4  -2  -4  -3  -2  -1  24  18  12   6  20  15  10
#> N  -9 -12  -9  -6  -6  -8  -6  -4  -3  -4  -3  -2  18  24  18  12  15  20  15
#> O  -6  -9 -12  -9  -4  -6  -8  -6  -2  -3  -4  -3  12  18  24  18  10  15  20
#> P  -3  -6  -9 -12  -2  -4  -6  -8  -1  -2  -3  -4   6  12  18  24   5  10  15
#> Q -16 -12  -8  -4 -12  -9  -6  -3  -8  -6  -4  -2  20  15  10   5  24  18  12
#> R -12 -16 -12  -8  -9 -12  -9  -6  -6  -8  -6  -4  15  20  15  10  18  24  18
#> S  -8 -12 -16 -12  -6  -9 -12  -9  -4  -6  -8  -6  10  15  20  15  12  18  24
#> T  -4  -8 -12 -16  -3  -6  -9 -12  -2  -4  -6  -8   5  10  15  20   6  12  18
#> U -20 -15 -10  -5 -16 -12  -8  -4 -12  -9  -6  -3  16  12   8   4  20  15  10
#> V -15 -20 -15 -10 -12 -16 -12  -8  -9 -12  -9  -6  12  16  12   8  15  20  15
#> W -10 -15 -20 -15  -8 -12 -16 -12  -6  -9 -12  -9   8  12  16  12  10  15  20
#> X  -5 -10 -15 -20  -4  -8 -12 -16  -3  -6  -9 -12   4   8  12  16   5  10  15
#>     T   U   V   W   X
#> A  -4 -20 -15 -10  -5
#> B  -8 -15 -20 -15 -10
#> C -12 -10 -15 -20 -15
#> D -16  -5 -10 -15 -20
#> E  -3 -16 -12  -8  -4
#> F  -6 -12 -16 -12  -8
#> G  -9  -8 -12 -16 -12
#> H -12  -4  -8 -12 -16
#> I  -2 -12  -9  -6  -3
#> J  -4  -9 -12  -9  -6
#> K  -6  -6  -9 -12  -9
#> L  -8  -3  -6  -9 -12
#> M   5  16  12   8   4
#> N  10  12  16  12   8
#> O  15   8  12  16  12
#> P  20   4   8  12  16
#> Q   6  20  15  10   5
#> R  12  15  20  15  10
#> S  18  10  15  20  15
#> T  24   5  10  15  20
#> U   5  24  18  12   6
#> V  10  18  24  18  12
#> W  15  12  18  24  18
#> X  20   6  12  18  24
#> 

seqs <- sh_extract_seqs(x$dt)

#> Task: extract specified-size sequences with input 'len' and 'step' under segment size cutoff.

#> Total segments is 467 and 343 of them with length >=1e+07

#> Fraction: 0.734

seqs

#> $keep
#> [1] "CFBFC" "FBFCG" "GJGKH" "HCFBF" "JGJGK" "OFOGO"
#> 
#> $drop
#>  [1] "B"    "BH"   "CFC"  "CHB"  "EK"   "EL"   "F"    "G"    "GJFK" "H"   
#> [11] "HBBC" "I"    "J"    "JK"   "JL"   "JNK"  "K"    "KGKG" "KHJ"  "KJ"  
#> [21] "KN"   "L"    "LJ"   "N"    "O"    "P"    "PSO"  "R"    "S"    "SKG" 
#> [31] "T"    "TWKS" "V"    "VW"   "W"    "WS"  
#> 

mat <- sh_get_score_matrix(seqs$keep, x$mat, verbose = TRUE)

#> Task: score equal-size paired strings with substitution matrix.

#> Handling index pair (1,1).
#> Handling index pair (2,1).
#> Handling index pair (2,2).
#> Handling index pair (3,1).
#> Handling index pair (3,2).
#> Handling index pair (3,3).
#> Handling index pair (4,1).
#> Handling index pair (4,2).
#> Handling index pair (4,3).
#> Handling index pair (4,4).
#> Handling index pair (5,1).
#> Handling index pair (5,2).
#> Handling index pair (5,3).
#> Handling index pair (5,4).
#> Handling index pair (5,5).
#> Handling index pair (6,1).
#> Handling index pair (6,2).
#> Handling index pair (6,3).
#> Handling index pair (6,4).
#> Handling index pair (6,5).
#> Handling index pair (6,6).

mat

#>       CFBFC FBFCG GJGKH HCFBF JGJGK OFOGO
#> CFBFC   120    90    43    85    14     9
#> FBFCG    90   120    38    90    23    20
#> GJGKH    43    38   120    36   -16   -30
#> HCFBF    85    90    36   120    26    12
#> JGJGK    14    23   -16    26   120    32
#> OFOGO     9    20   -30    12    32   120

mat2 <- sh_get_score_matrix(seqs$keep, x$mat, dislike = TRUE)

#> Task: score equal-size paired strings with substitution matrix.

identical(mat2, 120L - mat)

#> [1] TRUE

mat_b <- sh_get_score_matrix(seqs$keep, x$mat, block_size = 2L)

#> Task: score equal-size paired strings with substitution matrix.

## block1 represents the first 2 sequences
## block2 represents the 3rd, 4th sequences
## ...
mat_b

#>        block1 block2 block3
#> block1    105     64     17
#> block2     64     78     -2
#> block3     17     -2     76

mat_c <- sh_get_score_matrix(seqs$keep)

#> Task: score paired strings with longest common substring method.

#> Final score = 2^(LCS length - 1), for dislike=TRUE, max score is used to substract.

mat_c

#>       CFBFC FBFCG GJGKH HCFBF JGJGK OFOGO
#> CFBFC    16     8     0     8     0     1
#> FBFCG     8    16     1     4     1     1
#> GJGKH     0     1    16     1     8     1
#> HCFBF     8     4     1    16     0     1
#> JGJGK     0     1     8     0    16     1
#> OFOGO     1     1     1     1     1    16

mat_d <- sh_get_score_matrix(seqs$keep, dislike = TRUE)

#> Task: score paired strings with longest common substring method.

#> Final score = 2^(LCS length - 1), for dislike=TRUE, max score is used to substract.

mat_d

#>       CFBFC FBFCG GJGKH HCFBF JGJGK OFOGO
#> CFBFC     0     8    16     8    16    15
#> FBFCG     8     0    15    12    15    15
#> GJGKH    16    15     0    15     8    15
#> HCFBF     8    12    15     0    16    15
#> JGJGK    16    15     8    16     0    15
#> OFOGO    15    15    15    15    15     0

# \donttest{
if (require("doParallel")) {
  mock_seqs <- sapply(1:10000, function(x) {
    paste(sample(LETTERS[1:24], 5, replace = TRUE), collapse = "")
  })

  system.time(
    y1 <- sh_get_score_matrix(mock_seqs, x$mat, cores = 1)
  )

  system.time(
    y2 <- sh_get_score_matrix(mock_seqs, x$mat, cores = 2)
  )

  all.equal(y1, y2)
}

#> Loading required package: doParallel

#> Loading required package: foreach

#> Loading required package: iterators

#> Loading required package: parallel

#> Task: score equal-size paired strings with substitution matrix.

#> Task: score equal-size paired strings with substitution matrix.

#> [1] TRUE

# }